requirements.txt
streamlit==1.25.0
transformers==4.38.1
torch==2.0.1
pydub==0.25.1
yt-dlp==2025.04.01
matplotlib==3.8.0
pandas==2.1.0
python-docx==0.8.11
beautifulsoup4==4.12.2
numpy==1.26.0
accelerate
requests
pydub
